+++ NEWS.md
@@ -3,6 +3,8 @@
 * Reorders arguments for `decision` so that auto decisions can be specified without explicity
   referring to arguments
 
+* Finished coding `subroutine()` function and included package tests
+
 # NMproject 0.6.7
 
 * Added ability to use NMproject-specific code completion snippets `setup_code_completion()`.